net: Tell the user about a progress when opening the network session
I'm not sure whether I like this approach or not -- conceptually, the "waiting
for QNetworkSession to come up" is *not* a connection state, at least not with
the current Trojita's IMAP architecture where the IMAP code does not even open a
network connection until the NetworkWatcher tells it that the network is
available.

I guess it might be better to modify the GUI to show a clear and concise
indicator of the connectivity status which is separate from the IMAP state
machine altogether. Let's do that in another patch.

REVIEW: 116833