Fix missing inline copyright header for QMF stuff
It seems that back in 2009 when moving pieces of code between separate files I
forgot to add a proper copyright notice to the target files. Some stuff
(low-level encoding of RFC2047/quoted-printable/... bits) was imported from the
Qt Messaging Framework without copying an appropriate notice to the file header.
This is a mistake, and I hereby apologize for that. The top-level LICENSE file
already lists this contribution.

Thanks to Valério Valério (VDVsx on IRC) for bringing this to my attention.