Installed into system locations, trojita can't be started
This fixes not looking in standard data locations for the installed qml
files. To fix this it first checks if main.qml exists in the current
source tree, if not then we get a list of standard data locations and
try to find main.qml and throw qFatal if can't be found.

BUG: 332639
REVIEW: 117109

Conflicts:
	src/Ubuntu/main.cpp