Ubuntu: remove support for passing paths to QML files
There's no need to do this. The app will always load the QML from a path
relative to the application's own location (which handles the case for running
from inside the build directory). If that is not available, systemwide locations
are consulted as a fallback.

I just don't see any use case behind having to pass paths here.