Ubuntu: improved loading of error dialogs
I eventually figured out that if you passed the dialog as a URL you can use the
caller parameter (which in our case is the appWindow) which drives the dialogs
placement.

The third parameter can hold a series of properties configuring the popup to be
opened. This can be any property defined by the popup.

REVIEW: 117357