Ubuntu: Improved list item and fontsizes on message list view
The fontsizes are now using the FontUtils.sizeToPixels. It was
previously using
units.gu() which should only be used for spacing, anchors, margins etc.

Each List item was being rendered inside a fixed size rectangle, since a
listitem sets it's size based on
it's contents, there is no need to have fixed values for this but rather
use the size of the list item

REVIEW: 117371