Add support for the x-test language
Since we're using Qt's facilities for determining what language to use (based on
the technology we're using, we're a Qt application, not a KDE one), we end up
with examining the QLocale::system(). However, its return value is actually
sanitized by the C library (?), i.e. trying to use either LANG or LC_ALL set to
x-test or x_test does not work at all.

Rather than inventing a custom scheme, it's easier to just follow the KDE's
conventions for env vars, and this patch makes it follow the KDE_LANG env
variable when asked for the x-test language. We still follow the systemwide
settings, and the Qt's own locale is not changed (so you might still get
"untranslated" messages in Trojita in case they're provided by Qt itself).

REVIEW: 117429