Ubuntu: adjust the application name to play with the app confinement
QSettings uses the application's name and organization name to determine the
file path for storing config and cache data.  Because Ubuntu's app confinement
only allows access to certain pre-defined paths, we need to force the
application name and organization name to be something that will force QSettings
to use these allowed paths.

This has a pretty big disadvantage that the Ubuntu-specific version will not be
able to share data with the desktop one, unfortunately.

REVIEW: 117439