IMAP: relax quoted-string parsing to allow for Groupwise's escaped parentheses
As per the bugreport, Groupwise does not conform to RFC3501 and sends
parentheses ("(" and ")") as escaped characters within the quoted strings. The
RFC 3501 is pretty clear in only allowing escaped quotes and backslashes.

This change let's Trojita process these broken responses. It is a tough call to
tell what to do, but interpretting these by ignoring the backslashes is probably
the lesser of the two evils.

BUG: 334456
REVIEW: 118033