Get rid of the duplicated FLagsOperation
The QML is rather hostile towards working with enums which come from the C++
code *and* are not already wrapped in a QObject-deriving class. Instead of
duplicating the existing FlagsOperation enum partially, it seems that it is
cleaner to give up on the API design principles which ask for using enums
instead of plain bools for readability, and use a simple bool in this case.