Ubuntu: MessageListPage sorting
This patch makes the Listview load dynamically and force the view
to respect the user sorting prefs on each load. Triggering a
timer while the listview is populating, this benefits in that
we only sort the view once per view instead of using rowsInserted or like,
and reduce unneeded calls to.

The user can also switch sort order from the toolbar and if changed
the user sort order preference will be respected when going between
mailbox's.

navigation-menu.svg is part of the Ubuntu-mobile-icons package
available from http://packages.ubuntu.com/trusty/all/ubuntu-mobile-icons
and is available under the CC-BY-SA 3.0. Copyright by 2004-2012,
Canonical Ltd.

REVIEW: 118181