Reset auto-reconnect cycle when user explicitly changes network preference
Previously, a network error message box would be deleted and reinitialised
only when Trojita got a working connection to the server after a cycle
of reconnect attempts.

Now, whenever the user changes the network preferences, the
Imap::Model::NetworkWatcher apart from resetting the timeouts also emits a
signal that triggers deletion and reinitialisation of the error message
box at the GUI layer. This allows Trojita to treat a user's manual attempt
to re-connect as a new try-fail-reconnect cycle and throws a message pop-
up on a network error (even if the error string hasn't changed since last
time).

Use-Case: Error in network; Trojita shows an error message and starts a
series of auto reconnect attempts. User wishes to stop the attempts and
chooses "Offline" from the Network Access menu item. Attempts stop. After
a while, user wishes to try connecting again and chooses "Free Access".
If attempt to connect terminates in a network error, Trojita throws a
fresh new error message and starts a new auto-reconnect cycle (even if the
error string is the same as what it before the explicit switch to "offline").

REVIEW: 118434