Use QByteArray for various IMAP-related paths
Using QString doesn't really make sense in this context because the data we
operate on are absolutely not Unicode.

The change is rather intrusive, though, so I hope that I managed to catch all of
the instances where the drag-and-drop operated on QStrings. Maybe we should
switch to regular URLs here.