ComposeWidget: enable/disable the correct splitter
By using "splitter", the "mailText" text area was not enabled or
disabled. By using "verticalSplitter", this does happen so all widgets
are disabled when a message is being sent.

The (relevant part of the) hierarchy of the UI components looks like
this and explains why it solves the issue:

QSplitter verticalSplitter
---> QSplitter splitter
---> ComposerTextEdit mailText

REVIEW: 119398