tests: Add support for checking the client's output against a regular expression
v2: check that the pattern is not a multiline thing as that won't work with a
QRegExp, unfortunately.

v3: let the test fail with a proper error message when the Trojita-generated
data are wrong. Asserting here makes no sense whatsoever. Thanks to Stephan
Platz for his review.

Also adding some comments to better explain what's going on.