Support decrypting PGP/MIME messages
This commit is Jan's refactoring of Stephan's code contributed during
the GSoC 2014.

GnuPG access is abstracted through the QCA library (which calls out to
the `gpg` process, which is a bit ugly, but hey, it works).

No password prompts are apparently necessary. In the 2.x version of the
gpg tool (including 2.0.x, not just the "very-new" 2.1.x), passphrase
handling is being done by the agent, and the pinentry application is a
mandatory component.

Regarding the test suite -- the /dev/urandom faking is needed for
pre-2.1 GnuPG. Without that hack, GnuPG just insists on reading stuff
from /dev/random. I got bored after five minutes, and I can imagine that
the CI VMs won't have more entropy than this laptop. On the plus side,
the tests are now fully self-contained, the do not clobber the source
tree, and are pretty fast anyway.

Future work:
- add support to Composer so that it can access these fake parts
- support nested encryption
- support signed messages
- make this modular to support building this as a plugin
- extend QCA so that it informs us about additional metadata

Change-Id: Ifedb045dfc7f0f8a81d032a83f5eea78038aa578