searching message: scroll to current match
Because we're using the QWebView inside a QScrollArea container,
the attempts to scroll the QWebView itself have no direct effect.
Therefore we temporarily shrink the page viewport to the message
viewport (ie. the size it can cover at max), then perform the search,
store the scrollPosition, restore the page viewport and finally
scroll the messageview to the gathered scrollPosition,
mapped to the message (ie. usually the header)

Change-Id: I69d56dccd13342788d4211885ef6f9a885725324