Add proxy model for inserting/replacing random MIME subtrees
The new Model is a proxy model for the Imap::Mailbox::Model.
It provides a way to insert MessageParts into a message that
are not known to the IMAP server. This is useful for
displaying decrypted data from encrypted messages.

This is a joint work of Stephan Platz and Jan Kundrát.

Change-Id: I2b64ea455f4a29586a0fe3cbd5487546bd6891e5