IMAP: Add a proper state for waiting for a CLOSED response code
The CONN_STATE_SELECTING_WAIT_FOR_CLOSED describes a state where the
SELECT/EXAMINE command was already issued, but the IMAP state machine waits for
a CLOSED response code to know that the previously open mailbox was finally
closed, and any subsequent responses describe the new one.

v2: Don't call it "another mailbox" as this state will be used when opening a
mailbox for the first time as well.

Change-Id: I53f33c6667f14d154b5c4ba7a68cfffb8552f0ad