IMAP: Fix response routing while waiting for [CLOSED]
Turns out that the bug was that the KeepMailboxOpenTask already reseted its
isRunning flag, and hence decided to stop listening for these responses anymore.

We have to distinguish between the "not yet started" and "already finished"
states, if only for the resyncMailbox() to do the sane thing.

The reason on why one cannot actually use these responses in any way is that
there's a pretty big risk of introducing race conditions to the mix. Some of the
updates (like new message arrivals) absolutely require a followup action on our
side, such as discovering the UIDs, but we cannot do that at this time anymore
as another mailbox is already being selected. We *could* handle other updates
like VANISHED and FLAGS just fine, I guess, but then there would be that
annoying race condition when EXISTS arrives at first, we don't do anything about
it, and subsequently a FETCH updates the new arrival -- that would cause us to
scream out rather loudly. Another possible problem is handling only FLAGS
updates, but ignoring the EXISTS, etc.

Now I hope *this* is a reasonable final state.

BUG: 324761
Change-Id: Iaf9f75e7bfe6630f32af9cd6cea383e96f657598