Do not remember logs from old sessions indefinitely
Connection logs are typically most important for the current session, for that
special magic connection#0 which is the default log, and for a couple of the
recently closed connections to help with eventual troubleshooting. Tracking all
the past ones, however, doesn't make much sense and consumes a lot of resources
when the whole widget suddenly gets shown.

v2: Use a timer-based approach, not random skipping as per Thomas' suggestion.

Change-Id: Id98c390e0a82e7896fa4866ebb5c8372ef152b4f