Remove the Ubuntu touch version
The developers who were originally working on this decided that they want to
spend their time on their own fork instead. The fork is called Dekko for those
who are interested in helping them.

The reason for removeal of this code is that the Ubuntu-touch-specific version
is outdated, unamaintained and buggy. There is no point in shipping code which
nobody cares about, and thus it is going away now.

Change-Id: I2954b66b9066d7ee923bb9b9b67d59a7b854f4f1