Restore network connectivity after a brand new profile appears
There was a subtle bug -- if Trojita was already running and offline,
and a new network configuration got added (such as after plugging an
already-configured cellphone), the QNetworkSession was not getting
re-created with the new network configuration. This happened on X11 with
NetworkManager, where each QNetworkConfiguration apparently corresponds
to either a network device, or a WiFi AP, or a dial-up connection, etc.

Change-Id: I044625085fdb049c812eb0f178cca4a63356a686