IMAP: Handle missing UIDNEXT gracefully
Some ancient IMAP servers (I'm looking at you, Courier) do not report
UIDNEXT during SELECTs. This is of course suboptimal as it forces
Trojita to use a very inefficient syncing algorithm, but we should at
least prevent expensive trashing of the whole mailbox content, including
caches and the in-memory structure.

Luke-jr on IRC reported that when this happens on a big mailbox, the
ThreadingMsgListModel performs an extremely costly operation in its
pruneTree(). This is related to the way how we handle deletions at that
layer, by setting the internal ptr to 0. These messages are then pruned
one-by-one, and each of these deletes leads to a subsequent fixup of all
its siblings. An actual reported runtime was on the order of 45s with a
pretty beefy CPU. This patch does not fix the root cause of that
slowness (inefficient handleRowsDeleted/pruneTree), but it actually
prevents it from happening during the initial sync. It would be still
good to have an optimized threading model, but that is a topic for some
future patch.

I'm a little bit worried because this code doesn't raise any error if
the UID list changes in an incompatible way. It would be cool if it
checked the list of previously known UIDs and scream out loudly on
"impossible" changes, such as when a random UID gets inserted in the
middle or at the beginning.

Change-Id: I0501a8a9df969cef78b9f1bd12974fa2288b9129
CCBUG: 340241