Move STARTTLS persistence outside of the GUI
This is a logical change which should be implemented within the core
logic and not within a layer which is inherently UI-specific.

To put credit where credit is due, I found this while looking through
the public bug list for Dekko, the Ubuntu fork of Trojita. I'm not
saying "thanks for reporting", just stating what led me to fix this.

Change-Id: Id21273560b1e4511d38264da41bce0b448fd222b