Use iterators instead of operator[] for iterating a QVector
Both of these should be essentially free, but it happens that the debug
builds of Qt use boundary checking in the operator[], but not when using
iterators. Since there's absolutely no reason to use boundary checking
with this already-checked iteration, let's make sure it isn't needlessly
slow even in debug builds.

Change-Id: I9787c05f6489789bb1dad7bd2ea762436fc87fd4