Test tree node promoting while pruning the threading tree
This adds an explicit test which makes sure that the
ThreadingMsgListModel::pruneTree will do the right thing when it decides
to visit a three node whose preceding sibling was already pruned from
the tree. There might be some lurking asserts in there which would get
invalidated by that move of node offset fixup to happen for all nodes at
once.

Change-Id: I26027e1a79a8dcece7c4e23ece1959152a594b3a