Fix the RFC2047 decoder for "phrase production"
This is intended to fix a non-conformance when the decoder was used in
context of the "word in phrase" production as per RFC2047 section 5 (3).
That section makes it clear that e.g. parentheses are not allowed to be
sent unescaped in these contexts; that's because e.g. parentheses
indicate a RFC5322-style comment.

This is btw what Gerrit creates, but there's little that we could do
(besides fixing Gerrit, which I'll do shortly).

Change-Id: Ic0d3dff8408da3e28a34eadee4335a6c63c0525e