HTML: Accept quote marks interleaved with spaces in non-flowed mode
Since df8bff6e94273e5a6fb6bd3c018606f85b66dc6f, we employed a strict
algorithm which would ignore any quote level indication after a space
gets hit. While this is the right thing to do when parsing
format=flowed, it turns out that some MUAs out there not only have
user-configurable quote marks, but also default to *always* inserting
spaces between them -- KMail is one example, see the "KMail's quoting of
text in replies and format=flowed " thread at [1].

I believe that format=flowed is ultimately the only reasonable
specification of how to do quotes, and I would like to have a word with
people who think that replying with e.g. "|" as a quote separator. I am
not willing to go as far as to support *these* crazy separators, but in
order to better support these legacy clients which at least know ">",
but throw in an occasional space, this patch changes the code so that it
again recognizes this as a quote separator.

There's of course a downside to this as well -- older versions of
mailman would happily strip out format=flowed designation from the MIME
parts, and this means that Trojita will treat a line which is correctly
escaped in conformance with RFC3676, but which went through old mailman
by accident, and happens to start with an actual ">" as its content (and
not quote separator), in a wrong way.

[1] http://web.archiveorange.com/archive/v/vFLZWeVHCaFhIXmLa4oL

Change-Id: Id433b7f2a34e5f2f3f7f7d7ebccd78572e4db4fa