Fix: broken personal signature
With a single account registered in Trojitá the personal signature of the
user, properly configured in settings window, won't be presented in the 
end of text block in a *new email*, it works well just in replies. 

It occurs because the function responsible for this action, 
slotUpdateSignature(), is connected to the combobox item changed SIGNAL, 
in other words, if the user doesn't have any another account to emit the 
signal required then the function never will be called. As it is explicit 
called in setResponseData() function the signature appears only in 
replies.

The solution was to explicit call slotUpdateSignature() in the widget 
constructor, thus every time compose widget is created the signature is 
updated in the text block.  

Change-Id: Ia6b678a543dde3e41be8e74646ee69cf9a6c699f