Fix: unchecked condition crashes Trojitá.
If the "move down" button is clicked in the Settings dialog with no
item selection in the email list, Trojitá crashes.

It happens because the button is already enabled when the dialog opens,
with no selected item in the email list. When clicked, the value -1 is
propagated until Qt tries to change the item's position. Leading to a
segmentation fault.

The solution was to check the item selection: if no item is selected
(value -1) the button is disabled. Thus, this new condition aside those
already present, will disable the button when no item is selected and
when there is only one item.

Change-Id: Ic1d572799744d648be49bb036bb2ef876defb352