Prevent dereferencing of invalid iterator
This cannot happen in real world (the parser will send at least some
event during the process of logging out), but it's nonetheless a good
idea to be defensive from this context. It's quite possible that a
future change could add support for limiting the number of events which
are logged, and at that point we would have a problem. So better safe
than sorry here -- it's bad to rely on an assumption which is vague at
best.

Change-Id: I3ddfb4bed312f2e9d5e64588f9e5a5dca81d7d91