Report errors for message parts which failed to load
As far as I know, the code for TreeItem::isUnavailable has remained the
same since commit 92e0eb55695a05707b13a5b9efdb5b0c7c34ad4a. The way I
see it, it has always been buggy; if any TreeItem manages to get into
the UNAVAILABLE state, the only way of re-triggering its
download/fetch/whatever is through resetting the connection in some way.
That's the whole point of that state; it's an error, and the reason
behind that error could very well be that a connection went offline.

Now, stuff will be different when we support retrying of fetches and
when we won't rebuild the whole tree upon reconnect, but even then the
semantic of the UNAVAILABLE state seems to better match the meaning
introduced with this patch and not the former state.

Change-Id: Ie3ebe4567488a437264839b43acf13614e4072b8