Show unrecognized MIME parts as the default MIME part
There are at least two real-world classes of e-mail which fits this
heuristics:

- Outlook meeting invitations which come with an ICS file as a third
  item in multipart/alternative
- Thunderbird's inline forwards of certain HTML messages which come with
  an embedded PNG image. These are created as multipart/alternative
  where the first part is text/plain and the second part is a
  multipart/related containing a text/html and an image/png.

Especially the second case is very important because in that particular
case, the plaintext counterpart of the HTML message didn't contain any
indication of that attached image.

An unfortunate side effect of this is that the plaintext effectively
stops being preferred on some e-mails which contain any images referenced
from the HTML part (except when using data: URLs). This appears to be a
non-trivial problem to solve in a reasonable way :(. Fortunately, most
of the HTML spam I receive at least are:

 multipart/mixed
   multipart/alternative
     text/plain
     text/html
   image/jpeg

This patch effectively prefers disobeying user's preferences for
plaintext messages in cases where the HTML part *appears* to contain
some information which might be missing from the plaintext equivalent:

 multipart/alternative
   text/plain
   multipart/related
     text/html
     image/jpeg

I guess we'll see how well this works in real world.

Change-Id: I52bcade121a3afee4ab1301ebb89b8bd9407845a