Fix drag and drop of messages on Qt 5.4
In qtbase.git's commit 3ea0020d3b74560cd918af88b911781aceeceaef the
QAbstractItemModel now forwards calls to supportedDragActions to the
source model. Previously, it called the proxy's own
supportedDropActions. This means that all proxies had to effectively
return a unification of both drag and drop actions from
supportedDropActions, otherwise the drag functionality was limited.

Thanks to Thomas Lübking for identifying a root cause and to Dimitrios
Glentadakis for reporting this bug.

BUG: 342565
Change-Id: Icbc02dab9d091e315636e96bff1b8f63f544e856