Configure build type, optimizations and warnings
The motivation is to cut the number of supported build profiles a bit.
Stuff is now controlled through CMAKE_BUILD_TYPE. Optimizations are
active all the time (you can still override CXXFLAGS to change behavior),
and warnings are disabled on MSVC because we were told earlier that
their diagnostics leaves much to be desired. The -Werror is now merged
into the CMAKE_BUILD_TYPE as well for simplicity.

Change-Id: Ie535593d71052a6adce72ea7c3850575d1ab0ff3