Root mailbox should always correspond to an invalid QModelIndex
The root of all mailboxes (i.e. that special mailbox which is not
actually a mailbox, but a parent of everything else, including INBOX
etc) is supposed to be represented by an invalid QModelIndex. Turns out
that the code would sometimes represent that mailbox as a valid index
with row() == 0, column() == 0, internalPointer() == model->m_mailboxes.

IMHO there's a pretty big risk of this index leaking to the MVC users,
and given that there already is some infrastructure for making sure that
invalid QModelIndexes are handled as a special case, let's make sure
that this handling is applied consistently.

The root mailbox is not actually a mailbox; the only thing which has
common with "mailbox" is that it can be used for LISTing its children,
for fetching stuff from cache etc. Maybe it would make some sense to
special-case this, but that would require reimplementing the
load-children-from-cache and the entire ListChildMailboxesTask, so I
guess this is a slightly better alternative.

Change-Id: Ib52e102eb0be20ba788bf821226c7e1c32ddf5ea