Remove duplicate call to setFetchStatus
This is already unconditionally handled by the TreeItem::setChildren, so
there's ansolutely no point in calling that method again. Well, even the
comment right above that call says that.

I wonder if this implicit setFetchStatus is a good design, but stuff has
worked well for years, so I'm not going to change it right now (despite
the fact that there's at least one place which explicitly undoes this).

Change-Id: I0c4fb63329ad0cff74b63489fe231449f008d992