Don't LIST when it's already in progress
This prevents a duplicate LIST during application startup. There's no
associated test for this because it depends on what the GUI is doing.

The error which led to this patch was that the GUI code called something
that resulted in rowCount() very early in the startup, and Model's
switchover from offline mode to online enqueued another LIST even before
the old LIST got a chance to run.

The fix is to make sure that LIST cleans its mess when the command fails
for some reason, and that while a LIST is pending, no other LISTS are
going to be enqueued. I believe that this is a safe thing to do, because
e.g. a network error results in switching to offline, and hence killing
all tasks.

This has a very mild relation to bug 342808 because it makes the
situation a bit more manageable.

CCBUG: 342808
Change-Id: I8c06298bc209cc10b2b7c41bc60453b0f63c90d2